userpage, upload page and thumbnail worked on

+++ Minerva/base/views.py
@@ -1,14 +1,22 @@
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from  .forms import *
 from .models import *
+from .utils import *
 # Create your views here.
 
 def home(request):
     materials = Material.objects.all()
     return render(request, "base/home.html",{"materials": materials})
+
+def user(request):
+    return render(request, "base/user.html")
+    # user = User.objects.get(id = pk)
+    # materials = Material.objects.get(school = user.school, year = user.year, semester = user.semester).order_by("rating")
+    # return render(request, "base/user.html",{"materials": materials})
 
 def loginPage(request):
     page = "login"
@@ -62,10 +70,29 @@ def registerPage(request):
 def upload(request):
     if request.method == 'POST':
         material = MaterialForm(request.POST, request.FILES)
-  
+        actual = request.FILES.get('file')
+        file_type = fileTypeFinder(actual)
         if material.is_valid():
-            material.save()
-            return redirect('home')
+            mater = Material.objects.create(
+                host = request.user,
+                name = request.POST.get("name"),
+                author = request.POST.get("author"),
+                description = request.POST.get("description"),
+                course = Course.objects.get(id = 1),
+                file = actual,
+                type = file_type,
+            )
+            matrl = Material.objects.get(id = mater.id)
+            matrl.thumbnail = thumbnailer(mater.file)
+            matrl.type = "video"
+            return HttpResponse(matrl.type)
+            materl = Material.objects.get(id = material.id)
+            materl.thumbnail = thumbnailer(actual)
+            return redirect("home")
+
+        # if material.is_valid():
+        #     material.save()
+        #     return redirect('home')
     else:
         material = MaterialForm()
     return render(request, 'base/upload.html', {'material' : material})
@@ -73,6 +100,8 @@ def upload(request):
 
 def deleteFile(request,pk):
     material = Material.objects.get(id = pk)
+    if request.user != material.host:
+        return HttpResponse("You aren't allowed here mate")
     if request.method == 'POST':
         material.delete()
         return redirect('home')
